+++ scilab/modules/fileio/src/c/mputi.c
@@ -46,12 +46,12 @@ void C2F(mputi) (integer *fd,integer *res,integer *n,char type[],integer *ierr)
   ft_t ft;
   char *RES_c;
   unsigned char *RES_uc;
-  unsigned long *RES_ul;
+  uint32_t *RES_ul;
   unsigned short *RES_us;
 
   RES_c=(char *)res;
   RES_uc=(unsigned char *)res;
-  RES_ul=(unsigned long *)res;
+  RES_ul=(uint32_t *)res;
   RES_us=(unsigned short *)res;
 
   fa = GetFileOpenedInScilab(*fd);
@@ -72,7 +72,7 @@ void C2F(mputi) (integer *fd,integer *res,integer *n,char type[],integer *ierr)
 	case 'l' : 
 	  swap=SWAP(type,fd);
 	  if(swap<0) {*ierr=1;return;}
-	  MPUTI(long,swapl);
+	  MPUTI(int32_t,swapl);
 	  break;
 	case 's' : 
 	  swap=SWAP(type,fd);
@@ -103,8 +103,8 @@ void C2F(mputi) (integer *fd,integer *res,integer *n,char type[],integer *ierr)
 		     68000/SPARC style. */
 		  for ( i=0; i< *n; i++) 
 		    {
-		      unsigned long val;
-		      val =(unsigned long) *RES_ul++;
+		      uint32_t val;
+		      val =(uint32_t) *RES_ul++;
 		      wblong(ft, val);
 		    }
 		  break;
@@ -131,8 +131,8 @@ void C2F(mputi) (integer *fd,integer *res,integer *n,char type[],integer *ierr)
 		     VAX/386 style.*/
 		  for ( i=0; i< *n; i++) 
 		      {
-			unsigned long val;
-			val =(unsigned long) *RES_ul++;
+			uint32_t val;
+			val =(uint32_t) *RES_ul++;
 			wllong(ft, val);
 		      }
 		  break;
@@ -147,7 +147,7 @@ void C2F(mputi) (integer *fd,integer *res,integer *n,char type[],integer *ierr)
 		    }
 		  break;
 		default: 
-		  MPUTI(unsigned long,swapl);
+		  MPUTI(uint32_t,swapl);
 		  break;
 		}
 	      break;
